+++ python/mirage/workspace/names.py
@@ -32,6 +32,36 @@ UNSUPPORTED_BUILTINS = frozenset({
 
 NAMESPACE_COMMANDS = frozenset({"ln", "readlink"})
 
+# bash reserved words that mirage's grammar implements. The parser, not
+# the executor, consumes them, so they never reach route; `type` reports
+# them and the CLI registry refuses them as head words. bash's `time`
+# and `coproc` are left out on purpose: mirage implements neither
+# construct, so a line starting with one reports `command not found`,
+# and `type` may not contradict what dispatch does. Add a word back
+# when its construct lands.
+KEYWORDS = frozenset({
+    "if",
+    "then",
+    "else",
+    "elif",
+    "fi",
+    "case",
+    "esac",
+    "for",
+    "select",
+    "while",
+    "until",
+    "do",
+    "done",
+    "in",
+    "function",
+    "{",
+    "}",
+    "!",
+    "[[",
+    "]]",
+})
+
 # ShellBuiltin subset handled through the job table in the executor.
 JOB_BUILTINS = frozenset({"wait", "fg", "kill", "jobs", "ps"})
