+++ typescript/packages/core/src/workspace/executor/builtins/index.ts
@@ -49,5 +49,5 @@ export { handleTest } from './condition/index.ts'
 export { handleTimeout } from './timeout.ts'
 export { handleXargs } from './xargs.ts'
 export { handleCommandBuiltin } from './command.ts'
-export { handleType, handleWhich } from './lookup.ts'
+export { handleType, handleWhich } from './lookup/index.ts'
 export { handleEcho, handlePrintf } from './text.ts'
