+++ typescript/packages/core/src/workspace/executor/builtins/lookup/index.ts
@@ -1,0 +1,18 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+// The package's public surface: what other packages consume. Inside
+// the package, and in its tests, the modules are imported directly.
+export { classify, describe } from './classify.ts'
+export { handleType, handleWhich } from './handle.ts'
